Add menu to the stage screen
