+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/study_plan/delegate/LearningActivityTargetViewActionHandler.kt
@@ -3,8 +3,8 @@ package org.hyperskill.app.android.study_plan.delegate
 import androidx.fragment.app.Fragment
 import org.hyperskill.app.android.core.view.ui.navigation.requireRouter
 import org.hyperskill.app.android.projects_selection.list.navigation.ProjectSelectionListScreen
-import org.hyperskill.app.android.stage_implementation.view.dialog.UnsupportedStageBottomSheet
-import org.hyperskill.app.android.stage_implementation.view.navigation.StageImplementationScreen
+import org.hyperskill.app.android.stage_implementation.dialog.UnsupportedStageBottomSheet
+import org.hyperskill.app.android.stage_implementation.navigation.StageImplementationScreen
 import org.hyperskill.app.android.step.view.navigation.StepScreen
 import org.hyperskill.app.android.track_selection.list.navigation.TrackSelectionListScreen
 import org.hyperskill.app.learning_activities.presentation.model.LearningActivityTargetViewAction
