+++ androidHyperskillApp/src/main/res/menu/stage_menu.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+      xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/stageShare"
+        android:title="@string/step_menu_share_action"
+        app:showAsAction="never"/>
+
+    <item
+        android:id="@+id/stageFeedback"
+        android:title="@string/step_menu_report_action"
+        app:showAsAction="never"/>
+
+    <item
+        android:id="@+id/stageOpenInWeb"
+        android:title="@string/step_menu_open_in_web_action"
+        app:showAsAction="never"/>
+
+</menu>
